requirements.txt
pyTelegramBotAPI==4.29.1
diskcache==5.6.3
pytz==2025.2
httpx[http2]
Pillow==11.1.0